requirements.txt
aiogram==3.15.0
apscheduler==3.10.4
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
